refactor: Clean config files & reduce package's public api surface
